Add files via upload
7 ways to calculate number of steps.
The last one (integral based) fails really badly, which was a bit surprising, but I guess that is because of the nature of movement: when you move forward, turn around and don't "pendel back", the cumulative accelerations (forces) are positive (or negative).

The code starts to be a bit messy now, unnecessarily complex and could be improved in terms of transparency. But it works.